removed duplicate code from docker init.sh

+++ docker/addons/init.sh
@@ -125,11 +125,7 @@ if [[ ! -d ${HOST_VOLUME}/gitolite/ ]]; then
   cp -r /addons/gitolite/conf/* /home/${TENANT_USER}/gitolite-admin/conf/
   su - ${TENANT_USER} -c "cd /home/${TENANT_USER}/gitolite-admin; git add .; git commit -a -m 'Initial commit'; git push"
 fi
-# potential fix for auth keys
-su - ${TENANT_USER} -c "${HOST_VOLUME}/gitolite/bin/gitolite print-default-rc > ~/.gitolite.rc"
-su - ${TENANT_USER} -c 'sed -i "0,/# LOCAL_CODE/s//LOCAL_CODE/g" .gitolite.rc'
-su - ${TENANT_USER} -c "mkdir -p local/triggers"
-su - ${TENANT_USER} -c "echo -e '#!/bin/sh \n ln -sf /host_volume/ssh/authorized_keys ~/.ssh/' >> 'local/triggers/link-akf'"
+
 # activemq
 ln -s ${HOST_VOLUME}/activemq/data /opt/activemq/data
 if [[ ! -d ${HOST_VOLUME}/activemq ]]; then
@@ -163,7 +159,9 @@ fi
 
 if [[ ! -L /home/${TENANT_USER}/.ssh/authorized_keys ]]; then
   # put authorized_keys on the host volume to preserve it
-  su - ${TENANT_USER} -c "mv /home/${TENANT_USER}/.ssh/authorized_keys ${HOST_VOLUME}/ssh/"
+  if [[ -f /home/${TENANT_USER}/.ssh/authorized_keys ]]; then
+    su - ${TENANT_USER} -c "mv /home/${TENANT_USER}/.ssh/authorized_keys ${HOST_VOLUME}/ssh/"
+  fi
   ln -s ${HOST_VOLUME}/ssh/authorized_keys /home/${TENANT_USER}/.ssh/authorized_keys
 fi
